refactor - rename gateway to client
make explicit that the call made to the provided is through an http api
