+++ jobs/Backend/Task/ExchangeRateUpdater/Clients/ICzechNationalBankExchangeRateClient.cs
@@ -1,0 +1,8 @@
+﻿using ExchangeRateUpdater.Clients.Models;
+
+namespace ExchangeRateUpdater.Clients;
+
+public interface ICzechNationalBankExchangeRateClient
+{
+    CnbExchangeRates GetCurrentRates();
+}
