+++ jobs/Backend/Task/ExchangeRateUpdater/Clients/Models/CnbExchangeRates.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace ExchangeRateUpdater;
+namespace ExchangeRateUpdater.Clients.Models;
 
 public class CnbExchangeRates
 {
