--- jobs/Backend/Task/ExchangeRateUpdater/ICzechNationalBankExchangeRateGateway.cs
@@ -1,6 +1,0 @@
-﻿namespace ExchangeRateUpdater;
-
-public interface ICzechNationalBankExchangeRateGateway
-{
-    CnbExchangeRates GetCurrentRates();
-}
